requirements.txt
# --- Data Engineering ---
pandas
numpy
requests
beautifulsoup4
lxml
nba_api

# --- MLOps & Tracking ---
dvc[s3]
pyyaml
mlflow
python-dotenv
xgboost

# --- Entity Resolution ---
thefuzz
python-Levenshtein

# --- EDA & Visualization ---
jupyter
matplotlib
seaborn